fix ui

+++ templates/admin/users/userlist.php
@@ -10,19 +10,19 @@
                         <h3 class="font-weight-bold"> Danh Sách Thành Viên</h3>
                     </div>
                     <?php if (isset($_SESSION['failed'])) : ?>
-                    <div class="col-md-12">
-                        <div class="alert alert-danger mx-auto">
-                            <p><?= $_SESSION['failed'] ?></p>
+                        <div class="col-md-12">
+                            <div class="alert alert-danger mx-auto">
+                                <p><?= $_SESSION['failed'] ?></p>
+                            </div>
                         </div>
-                    </div>
                     <?php endif; ?>
 
                     <?php if (isset($_SESSION['success'])) : ?>
-                    <div class="col-md-12">
-                        <div class="alert alert-success mx-auto">
-                            <p><?= $_SESSION['success'] ?></p>
+                        <div class="col-md-12">
+                            <div class="alert alert-success mx-auto">
+                                <p><?= $_SESSION['success'] ?></p>
+                            </div>
                         </div>
-                    </div>
                     <?php endif; ?>
                     <table class="table table-bordered">
                         <thead>
@@ -55,17 +55,17 @@
                         $num = 1;
                         ?>
                         <?php foreach ($result as $users) : ?>
-                        <tbody>
-                            <tr>
-                                <th scope="row" class="text-center"><?= $num++ ?></th>
-                                <th><?= $users['username'] ?></th>
+                            <tbody>
+                                <tr>
+                                    <th scope="row" class="text-center"><?= $num++ ?></th>
+                                    <th><?= $users['username'] ?></th>
 
-                                <th><?= $users['firstName'] ?></th>
-                                <th><?= $users['lastName'] ?></th>
-                                <th><?= $users['phone'] ?></th>
-                                <th><?= $users['birthday'] ?></th>
-                                <th><?= $users['gender'] ?></th>
-                                <th> <?php
+                                    <th><?= $users['firstName'] ?></th>
+                                    <th><?= $users['lastName'] ?></th>
+                                    <th><?= $users['phone'] ?></th>
+                                    <th><?= $users['birthday'] ?></th>
+                                    <th><?= $users['gender'] ?></th>
+                                    <th> <?php
                                             if ($users['id_role'] == 1) {
                                                 echo 'Master';
                                             } else if ($users['id_role'] == 2) {
@@ -77,18 +77,16 @@
                                             }
 
                                             ?></th>
-                                <th><?= $users['created_at'] ?></th>
-                                <th><?= $users['updated_at'] ?></th>
-                                <th>
-                                    <a href="?action=fixuserId=<?= $users['id'] ?> "><i style="margin-left: 20px;"
-                                            class="fa fa-edit" aria-hidden="true"></i></a>
-                                    <a onclick="return confirm('Bạn có muốn xóa?')"
-                                        href="templates/admin/users/delete.php?deleteId=<?= $users['id'] ?>">
-                                        <i style="margin-left: 10px;" class="fa fa-trash" aria-hidden="true"></i></a>
+                                    <th><?= $users['created_at'] ?></th>
+                                    <th><?= $users['updated_at'] ?></th>
+                                    <th>
+                                        <a href="?action=fixuserId=<?= $users['id'] ?> "><i style="margin-left: 20px;" class="fa fa-edit" aria-hidden="true"></i></a>
+                                        <a onclick="return confirm('Bạn có muốn xóa?')" href="templates/admin/users/delete.php?deleteId=<?= $users['id'] ?>">
+                                            <i style="margin-left: 10px;" class="fa fa-trash" aria-hidden="true"></i></a>
 
-                                </th>
-                            </tr>
-                        </tbody>
+                                    </th>
+                                </tr>
+                            </tbody>
                         <?php endforeach; ?>
                     </table>
                 </div>
